[DATALAD] removed content
This removes the inputs/joe_and_lili subdataset (which contained
a sub-subdataset, inputs/joe_and_lili/source/EI_clustered_network).

Binder seems to do a deep (recursive) clone by default. I am removing
the subdataset to simplify the setup -- subdatasets per se are not
a problem, but EI_clustered_network was recorded in a commit which is
not available in the clone (not sure why), leading to errors.

We may want to revert this commit and fix the sub-subdataset, but for
now I want to explore other things.