Move Dockerfile and environment.yml to top level
It seems to me that Binder only picks up the environment definitions
if they are stored in the top level. AFAIK, if Dockerfile is present,
others will be ignored - but the environment.yml is referenced from
within the Dockerfile, so we need it too.

Let's see where it takes us.